requirements.txt
openai>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
supabase>=2.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
